snap.getFragFromBarcode: fix omission of decoding sequence names
Under Python 3, the sequence (chromosome) names provided to MACS2 were
retrieved as byte strings, then wrapped in `str()` before writing to disk.
This resulted in sequence names like "b'chr1'", as in:

b'chr14'      46787083        46787310        AAACATCGAAGGACACCATCAAGT
b'chr3'       56262063        56262116        AAACATCGAAGGACACCATCAAGT
b'chr2'       103894442       103894639       AAACATCGAAGGACACCATCAAGT

Tested this fix under Python 2 and 3.

(This duplicates a commit in the 'dump-frag-py3-fix' branch, but that commit
has merge conflicts due to code reformatting, and it's trivial to recreate.)